requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
obs-websocket-py==1.0
obsws-python==1.8.0
simpleaudio==1.0.4
pydub==0.25.1
websocket-client==1.8.0
Werkzeug==3.1.3
matplotlib==3.8.3
pandas==2.2.2
websockets==12.0
pygame==2.6.1
inputs==0.5
psutil==5.9.6
bleak==1.0.1
requests==2.32.4
google-generativeai==0.8.5
pycountry==24.6.1
Flask-SocketIO==5.5.1
Pillow==11.3.0
pydantic==2.11.7
numpy==1.26.4
scipy==1.16.1
python-dotenv==1.0.1
PyYAML==6.0.2
